Move UnsatisfiedInjectableReferenceTest and AmbiguousInjectableReferenceTest under appropriate package.
