+++ impl/src/main/java/org/jboss/cdi/tck/tests/full/lookup/injectionpoint/broken/reference/ambiguous/DerivedInjectedBean.java
@@ -14,19 +14,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.cdi.tck.tests.lookup.injectionpoint.broken.reference.ambiguous;
+package org.jboss.cdi.tck.tests.full.lookup.injectionpoint.broken.reference.ambiguous;
 
 import jakarta.enterprise.context.Dependent;
 
 @Dependent
-public class SimpleBean {
-    private InjectedBean injectedBean;
+public class DerivedInjectedBean extends InjectedBean {
 
-    public InjectedBean getInjectedBean() {
-        return injectedBean;
-    }
-
-    public void setInjectedBean(InjectedBean injectedBean) {
-        this.injectedBean = injectedBean;
-    }
 }
